feat: add saleOrder with docker-compose production

+++ app/model/order_model.go
@@ -2,7 +2,7 @@ package model
 
 type Order struct {
 	BaseModel
-	TotalPrice  int `json:"total_price" binding:"required"`
-	TotalAmount int `json:"total_amount" binding:"required"`
+	TotalPrice  float64 `json:"total_price" binding:"required"`
+	TotalAmount int     `json:"total_amount" binding:"required"`
 	OrderDetail []OrderDetail
 }
